@@ -30,23 +30,20 @@
     "node": ">=12"
   },
   "devDependencies": {
-    "@types/node": "14.14.19",
-    "@typescript-eslint/eslint-plugin": "4.11.1",
-    "@typescript-eslint/parser": "4.11.1",
+    "@types/node": "16.4.13",
+    "@typescript-eslint/eslint-plugin": "4.29.0",
+    "@typescript-eslint/parser": "4.29.0",
     "conventional-github-releaser": "3.1.5",
-    "dotenv": "8.2.0",
-    "eslint": "^7.17.0",
-    "eslint-config-prettier": "7.1.0",
-    "eslint-plugin-prettier": "3.3.0",
-    "homebridge": "1.1.7",
-    "homebridge-config-ui-x": "4.36.0",
-    "prettier": "^2.2.1",
-    "standard-version": "9.1.0",
-    "typescript": "4.1.3"
+    "dotenv": "10.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-prettier": "3.4.0",
+    "homebridge": "1.3.4",
+    "prettier": "^2.3.2",
+    "standard-version": "^9.3.1",
+    "typescript": "4.3.5"
   },
-  "dependencies": {
-    "somfy-synergy": "^1.4.0"
-  },
+  "dependencies": {},
   "files": [
     "CHANGELOG.md",
     "config.schema.json",
